user 디렉토리 user로 이름 변경
