+++ src/main/java/com/matzip/thread/security/provider/ApiAuthenticationProvider.java
@@ -3,7 +3,7 @@ package com.matzip.thread.security.provider;
 import com.matzip.thread.common.exception.ApiAuthenticationException;
 import com.matzip.thread.security.model.UserContext;
 import com.matzip.thread.security.token.ApiAuthenticationToken;
-import com.matzip.thread.users.domain.User;
+import com.matzip.thread.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.core.Authentication;
