+++ src/main/java/com/matzip/thread/user/adapter/out_/persistence/UserJpaRepository.java
@@ -1,0 +1,10 @@
+package com.matzip.thread.user.adapter.out_.persistence;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+interface UserJpaRepository extends JpaRepository<UserJpaEntity, Long> {
+
+    Optional<UserJpaEntity> findByUsername(String username);
+}
