+++ src/main/java/com/matzip/thread/user/domain/Role.java
@@ -1,0 +1,6 @@
+package com.matzip.thread.user.domain;
+
+public enum Role {
+    ROLE_ADMIN,
+    ROLE_USER
+}
