+++ src/main/java/com/matzip/thread/user/domain/User.java
@@ -1,0 +1,13 @@
+package com.matzip.thread.user.domain;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class User {
+    private final String username;
+    private final String nickname;
+    private final String password;
+    private final Role role;
+}
