--- src/main/java/com/matzip/thread/users/domain/Role.java
@@ -1,6 +1,0 @@
-package com.matzip.thread.users.domain;
-
-public enum Role {
-    ROLE_ADMIN,
-    ROLE_USER
-}
