+++ src/test/java/com/matzip/thread/security/SecurityTest.java
@@ -2,9 +2,9 @@ package com.matzip.thread.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.matzip.thread.security.model.SignInRequest;
-import com.matzip.thread.users.application.port.in.UserInPort;
-import com.matzip.thread.users.domain.Role;
-import com.matzip.thread.users.domain.User;
+import com.matzip.thread.user.application.port.in.UserInPort;
+import com.matzip.thread.user.domain.Role;
+import com.matzip.thread.user.domain.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.BDDMockito;
