+++ src/test/java/com/matzip/thread/user/domain/UserTest.java
@@ -1,0 +1,12 @@
+package com.matzip.thread.user.domain;
+
+import org.junit.jupiter.api.Test;
+
+class UserTest {
+
+    @Test
+    void test() {
+
+    }
+
+}
