--- src/test/java/com/matzip/thread/users/domain/UserTest.java
@@ -1,14 +1,0 @@
-package com.matzip.thread.users.domain;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserTest {
-
-    @Test
-    void test() {
-
-    }
-
-}
